Refactor: Move AttachToConsole to console package
