+++ pkg/virtctl/guestfs/BUILD.bazel
@@ -7,8 +7,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/pointer:go_default_library",
+        "//pkg/virtctl/console:go_default_library",
         "//pkg/virtctl/templates:go_default_library",
-        "//pkg/virtctl/utils:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
