+++ pkg/virtctl/utils/BUILD.bazel
@@ -2,11 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "test_utils.go",
-        "utils.go",
-    ],
+    srcs = ["test_utils.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virtctl/utils",
     visibility = ["//visibility:public"],
-    deps = ["//vendor/golang.org/x/term:go_default_library"],
 )
